+++ libs/react/src/__mocks__/sidebarMock.ts
@@ -40,6 +40,14 @@ export function getSidebarMock() {
           route: '/compliance',
         },
         {
+          key: 'documents_key',
+          icon: {
+            name: 'ColdDocumentsIcon',
+          },
+          label: 'Documents',
+          route: '/documents',
+        },
+        {
           key: 'actions_key',
           icon: {
             name: 'ColdActionsIcon',