+++ libs/react/src/components/application/routes/routes.tsx
@@ -1,4 +1,4 @@
-import { DashboardLayout } from '../../pages';
+import { DashboardLayout, DocumentUpload } from '../../pages';
 import { Route, Routes } from 'react-router-dom';
 import { Settings } from '../../pages';
 import { Home } from '../../pages';
@@ -32,7 +32,7 @@ export const ColdRoutes = () => {
               <Route path={'/settings'} element={<Settings />} />
               {ldFlags.showActions261 && ActionRoutes()}
               {ldFlags.showComplianceModule && ComplianceRoutes()}
-              <Route path="/compliance" element={<div className={'text-tc-primary'}>Pending...</div>} />
+              {ldFlags.showComplianceModule && <Route path="/documents" element={<DocumentUpload />} />}
               <Route path="*" element={<div className={'text-tc-primary'}>Pending...</div>} />
             </Route>
           </Route>