+++ libs/react/src/components/atoms/icons/coldIcon.tsx
@@ -20,6 +20,7 @@ import { ColdChevronUpIcon } from './coldChevronUpIcon';
 import { ColdChevronDownIcon } from './coldChevronDownIcon';
 import { ColdCheckIcon } from './coldCheckIcon';
 import { ColdComplianceIcon } from './coldComplianceIcon';
+import { ColdDocumentsIcon } from './coldDocumentsIcon';
 
 // TODO: Take the color attribute out of the IconProps
 export const ColdIcon = (props: IconProps) => {
@@ -61,6 +62,8 @@ export const ColdIcon = (props: IconProps) => {
       return <ColdCheckIcon {...props} />;
     case IconNames.ColdComplianceIcon:
       return <ColdComplianceIcon {...props} />;
+    case IconNames.ColdDocumentsIcon:
+      return <ColdDocumentsIcon {...props} />;
     default:
     case IconNames.ColdSettingsIcon:
       return <ColdSettingsIcon {...props} />;